Move middleware into Middlewares folder.
