+++ src/WebUI/Startup.cs
@@ -1,7 +1,3 @@
-using Codidact.Application;
-using Codidact.Infrastructure;
-using Codidact.Infrastructure.Persistence;
-
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.EntityFrameworkCore;
@@ -9,6 +5,11 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
+
+using Codidact.Application;
+using Codidact.Infrastructure;
+using Codidact.Infrastructure.Persistence;
+using Codidact.WebUI.Middlewares;
 
 namespace Codidact.WebUI
 {
